Optimization - added react based optimization (useMemo, memo, useCallback)
