+++ src/App.css
@@ -271,7 +271,7 @@ img{
   section{
     margin:180px 0;
     scroll-margin-top: 80px;
-    padding: 1rem 4rem;
+    padding: 1rem 2rem;
   }
 
   .cta{
