+++ src/components/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import "../css/Footer.css";
+import "../../styles/Footer.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faGithub, faLinkedinIn } from "@fortawesome/free-brands-svg-icons";
