+++ src/components/ui/SmallCard.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import { memo, ReactNode } from 'react';
 import CountUp from 'react-countup';
 
 interface SmallCardProps {
@@ -8,7 +8,7 @@ interface SmallCardProps {
   desc:string;
 }
 
-function SmallCard({section,stats, icon, desc}:SmallCardProps) {
+const SmallCard = memo(({section,stats, icon, desc}:SmallCardProps) => {
   return (
     <div className={`${section}__sCard`}>
       <div className="sCard__visual">
@@ -29,12 +29,6 @@ function SmallCard({section,stats, icon, desc}:SmallCardProps) {
       <p className="sCard__content">{desc}</p>
     </div>
   )
-}
-
-SmallCard.defaultProps = {
-  stats: 0,
-  icon: null,
-  desc: 'No description available' 
-}
+})
 
 export default SmallCard
