+++ src/components/ui/StackCard.tsx
@@ -1,8 +1,10 @@
+import { FC } from "react";
+
 interface StackCardProps {
   tech: string;
 }
 
-function StackCard({tech}: StackCardProps) {
+const StackCard: FC<StackCardProps> =({tech}: StackCardProps) => {
   return (
     <span className='stack__card'>{tech}</span>
   )
