+++ src/styles/animations/Stars.css
